10_construction_bom/requirements.txt
langchain
langchain-openai
langchain-community
pypdf
pandas
streamlit
python-dotenv
